Finished refactoring for now. Im done heree

+++ src/main/java/com/david/dev/portfolio_be/service/LLMChatUserService.java
@@ -1,10 +1,12 @@
 package com.david.dev.portfolio_be.service;
 
 import com.david.dev.portfolio_be.model.LLMChatUser;
+import com.david.dev.portfolio_be.model.dto.LLMChatUserDTO;
+import com.david.dev.portfolio_be.model.mapper.LLMChatUserMapper;
 import com.david.dev.portfolio_be.repository.LLMChatUserRepository;
 import org.springframework.stereotype.Service;
 
-import java.util.Collection;
+import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -17,20 +19,31 @@ public class LLMChatUserService {
         this.llmChatUserRepository = llmChatUserRepository;
     }
 
-    public Collection<LLMChatUser> getAllLLMChats() {
-        return llmChatUserRepository.findAllByOrderByUpdatedAtDesc();
+    public List<LLMChatUserDTO> getAllLLMChats() {
+        return llmChatUserRepository.findAllByOrderByUpdatedAtDesc().stream()
+                .map(LLMChatUserMapper::toDTO)
+                .toList();
     }
 
-    public Optional<LLMChatUser> getAllLLMChatsByChat(UUID id) {
-        return llmChatUserRepository.findById(id);
+    public LLMChatUserDTO getAllLLMChatsByChat(UUID id) {
+        LLMChatUser llmChatUser = llmChatUserRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("LLMChatUser not found with id: " + id));
+
+        return LLMChatUserMapper.toDTO(llmChatUser);
     }
 
-    public LLMChatUser createLLMChatUser(LLMChatUser llmChatUser){
-        return llmChatUserRepository.save(llmChatUser);
+    public LLMChatUserDTO createLLMChatUser(LLMChatUserDTO llmChatUserDto){
+        LLMChatUser llmChatUser = LLMChatUserMapper.toEntity(llmChatUserDto);
+
+        return LLMChatUserMapper.toDTO(llmChatUserRepository.save(llmChatUser));
     }
 
-    public void deleteLLMChat(UUID id) {
+    public boolean deleteLLMChat(UUID id) {
+        Optional<LLMChatUser> optional = llmChatUserRepository.findById(id);
+        if (optional.isEmpty()) return false;
+
         llmChatUserRepository.deleteById(id);
+        return true;
     }
 
 }
